Refactoring and add feat mail
