--- internal/domain/partner/partner_interface.go
@@ -1,1 +1,0 @@
-package partner
