+++ internal/helpers/email_valid.go
@@ -1,0 +1,8 @@
+package helpers
+
+import "net/mail"
+
+func Valid(email string) bool {
+	_, err := mail.ParseAddress(email)
+	return err == nil
+}
