+++ public/template/reset_email_template_simple.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Email Template</title>
+</head>
+<body>
+
+<h1>Verification code</h1>
+<p>Hi {{ .Username }},</p>
+<p>Here is your verification code: {{ .Code }}</p>
+
+<p>Thank you</p>
+<p>Best regards</p>
+<p>Team</p>
+
+</body>
+</html>
