--- utils/email_valid.go
@@ -1,8 +1,0 @@
-package utils
-
-import "net/mail"
-
-func Valid(email string) bool {
-	_, err := mail.ParseAddress(email)
-	return err == nil
-}
