--- utils/generate_uid.go
@@ -1,8 +1,0 @@
-package utils
-
-import "github.com/google/uuid"
-
-func NewUUIDString() string {
-	uid := uuid.New().String()
-	return uid
-}
