--- utils/hash_password.go
@@ -1,12 +1,0 @@
-package utils
-
-import "golang.org/x/crypto/bcrypt"
-
-func HashPassword(password string) (string, error) {
-	bytesPass, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	return string(bytesPass), err
-}
-
-func ComparePassword(hashedPassword, password string) error {
-	return bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
-}
