--- utils/response.go
@@ -1,10 +1,0 @@
-package utils
-
-import "github.com/gofiber/fiber/v2"
-
-func Response(c *fiber.Ctx, statusCode int, message string, data interface{}) error {
-	return c.Status(statusCode).JSON(fiber.Map{
-		"message": message,
-		"data":    data,
-	})
-}
